Backend/location-service/requirements.txt
flask>=3.0.0
grpcio>=1.71.0
grpcio-tools>=1.71.0
protobuf>=5.29.0
python-dotenv>=1.0.0